DEV: Initial commit after breaking down the img proc mods.
This is the initial commit in the new image arithmetic branch, which has
been created in order to simplify addition and merging of the compiled
image processing tools to skxray. This branch will focus exclusively on
image arithmetic and logical operations, tests of these functions,
demonstration of function use, as well as helper functions to a) keep
track of, and present data set attributes, and b) validate and adjust
attributes such as array dimensions if input arrays mismatch. User
friendly errors are also being integrated to aid the user in correcting
errors such as differences in dimensionality, or dtype.