DEV: Added helper function to check input dims and correct if mismatched
In my last meeting with Wah-Keat he asked the question of what occurs if
the two input arrays are mismatched in dimensions. Of course this raises
an error, and prevents image processing from continuing. This helper
function doesn't modify the input arrays, it simply evaluates array
dimensionality and returns the array dimensions that are required in
order for the input arrays to be compared.