DEV: Added test helper function to check for application of constant val
If a test array is being modified by applying a constant value (e.g.
multiplying all values in an array by 10) then this is a case where the
array dimension check is unnecessary. The added code evaluates the input
data to check for constant inputs. If one (or more) of the input sources
contains only a constant value, then no further processing occurs.