DEV: Added helper test to verify both inputs have same # of axes
The test has been added to the helper function _check_array_size. This
test checks for situations where a 2D image and a 3D image have been
specified as the two input arrays. In the event that something like this
happens, a warning message will be generated and the image processing
workflow will be stopped.