DEV: Added test to helper function to detect 1D arrays and gen warning
Since 1D arrays are atypical, but potentially valid image, or data
inputs, this test simply warns the user that the input arrays are
1-dimensional. Potential instances where this might occur include 1) The
user converts a VTK data set back into a numpy array but forgets to
resize the array. Since array data in VTK objects is stored as a 1D
string, conversion back to numpy requires the added step of resizing the
array to the proper dimensions.

It should also be noted that 1D input arrays will only survive the tests
if BOTH input arrays are 1D, or a constant is being applied to the 1D
    array.