TST: Added test functions for application of a constant to an array
Three tests are currently included in the test function:
test_apply_constant. They focus on 1) application of an int constant to
an int array, which works as expected, returning valid_operation=True,
resize_arrays=False, and corrected_dims=(shape of int array). 2) Tests
application of a Float constant to an Int array, which should generate a
TypeError. 3) Simply to round out the testing the third applies a Float
constant to a Float array, which should evaluate to valid_operation=True,
resize_arrays=False, and corrected_dims=(shape of int array).