TST: Added to documentation and added tst for arrays with eq dims.
The test evaluation for two arrays, of same dtype, and equal dimensions
has been added to the test function test_array_size_check(). The test
makes sure that if the input arrays are as they should be (e.g. same
dtype with equal shape and dimensions) then the helper function returns:
valid_operation=True, resize_arrays=False, and corrected_dims=(shape of
array_1). Since the arrays are as they should be, the key resize_arrays
will indicate to the array_resize helper function that no modification
is necessary and the arithmetic operation should commence normally.