DEV: Modified the constant value checker in _check_array_size.
Modified the constant value checker in _check_array_size to work as
intented. The test now evaluates whether either of the input objects are
a constant. The test sorts out which input is an array and which is a
constant. The test then confirms that both inputs are of the same type,
and if they are not then a TypeError is raised and advice for fixing the
problem is provided in the error message.